+++ DuoArtifacts/docs/release-notes.md
@@ -6,6 +6,7 @@
 
 * Added possible duos to character tooltips on the new run screen.
 * Fixed Drone Overclock (the Isaac-Drake duo) not working with Shield Drones as advertised.
+* Fixed HARRIER Protocol (the Peri-Riggs duo) not limiting Evade as advertised.
 
 ## 1.0.6
 Released 17 March 2024.